ICP-2 Weather HTML Submission
I named it as weather2 to be able to differentiate between the source code and the submitted code. It started out as the clone of the first, but more code has been added and modified extensively thereafter with references from the Array.html, Directions.html, Array.html, and localstorage.html as well.